looping v2 frontend working

+++ packages/builder/src/components/automation/AutomationBuilder/StepPanel.svelte
@@ -132,20 +132,22 @@
           {loopBlock ? `Stop Looping` : `Loop`}
         </ActionButton>
       {/if}
-      <ActionButton
-        quiet
-        noPadding
-        icon="trash"
-        on:click={async () => {
-          if (!blockRef) {
-            return
-          }
-          await automationStore.actions.deleteAutomationBlock(blockRef.pathTo)
-        }}
-      >
-        Delete
-      </ActionButton>
-      {#if $memoBlock && !isBranchStep($memoBlock)}
+      {#if !blockRef?.loopV2Child}
+        <ActionButton
+          quiet
+          noPadding
+          icon="trash"
+          on:click={async () => {
+            if (!blockRef) {
+              return
+            }
+            await automationStore.actions.deleteAutomationBlock(blockRef.pathTo)
+          }}
+        >
+          Delete
+        </ActionButton>
+      {/if}
+      {#if $memoBlock && !isBranchStep($memoBlock) && !blockRef?.loopV2Child}
         <ActionButton
           quiet
           noPadding
